chore: move mock files into root test folder
