+++ packages/bridge-controller/src/bridge-controller.test.ts
@@ -7,13 +7,13 @@ import BridgeController from './bridge-controller';
 import { DEFAULT_BRIDGE_CONTROLLER_STATE } from './constants';
 import { CHAIN_IDS } from './constants/chains';
 import { SWAPS_API_V2_BASE_URL } from './constants/swaps';
-import mockBridgeQuotesErc20Native from './test/mock-quotes-erc20-native.json';
-import mockBridgeQuotesNativeErc20Eth from './test/mock-quotes-native-erc20-eth.json';
-import mockBridgeQuotesNativeErc20 from './test/mock-quotes-native-erc20.json';
 import type { BridgeControllerMessenger, QuoteResponse } from './types';
 import { getBridgeApiBaseUrl } from './utils';
 import * as balanceUtils from './utils/balance';
 import * as fetchUtils from './utils/fetch';
+import mockBridgeQuotesErc20Native from '../../../tests/bridge-controller/mock-quotes-erc20-native.json';
+import mockBridgeQuotesNativeErc20Eth from '../../../tests/bridge-controller/mock-quotes-native-erc20-eth.json';
+import mockBridgeQuotesNativeErc20 from '../../../tests/bridge-controller/mock-quotes-native-erc20.json';
 import { flushPromises } from '../../../tests/helpers';
 
 const EMPTY_INIT_STATE = {
